requirements.txt
Flask
Pillow
python-dotenv
google-generativeai
streamlit
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
flask_socketio
google-genai